In SetFilter, filterParams.suppressSorting has no effect
The suppressSorting flag in SetFilter's filterParams should suppress sorting of the item list. It does sort them however, even if you pass suppressSorting = true in the filterParams. There is a local variable suppressSorting which is passed (unitialized) to the SetFilterModel instead of filterParams.suppressSorting. This change removes the local variable and uses filterParams.suppressSorting.